webgl / separate ShaderBuilder from style parsing
Also remove unused methods. Intended for clarity
